+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 * Logged requests were always shown as belonging to the Delivery API, even when the Preview API was used.
 * Fields not present in an Entry would lead to an error.
 * The Synchronization Manager's method `startSync` was type hinted to the wrong Query class.
+* API responses were not correctly logged.
 
 ## [0.6.2-beta](https://github.com/contentful/contentful.php/tree/0.6.2-beta) (2016-02-22)